general: refactoring a bit, removing the main process and stub exceptions, rephrasing README.md
