--- timewatch/timewatch.py
@@ -1,74 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import requests
-from bs4 import BeautifulSoup
-
-from .exceptions import TimeWatchLoginError, TimeWatchError
-
-SITE_URL = "http://checkin.timewatch.co.il/punch/"
-PUNCH_IN = r"כניסה"
-PUNCH_OUT = r"יציאה"
-
-HEADERS = {
-    'User-Agent': "Mozilla/5.0 (Windows NT 6.3; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/70.0.3538.102 Safari/537.36",
-    'Referer': "http://checkin.timewatch.co.il/punch/punch2.php"
-}
-
-
-def parse_html(html_doc):
-    return BeautifulSoup(html_doc, 'html.parser')
-
-
-def urljoin(*parts):
-    parts = map(lambda x: x.strip('/'), parts)
-    return '/'.join(parts)
-
-
-class TimeWatch(object):
-    def __init__(self, company, employee, password):
-        self.ix_employee = ''
-        self.company = company
-        self.employee = employee
-        self.password = password
-        self.session = requests.Session()
-
-    def login(self):
-        res = self.session.post(urljoin(SITE_URL, 'punch2.php'),
-                                data={
-                                    'comp': self.company,
-                                    'name': self.employee,
-                                    'pw': self.password
-                                }, allow_redirects=False, headers=HEADERS)
-
-        if res.status_code != 200:
-            raise TimeWatchLoginError(res.text)
-        parsed = parse_html(res.text)
-        self.ix_employee = \
-            parsed.find('input', {'id': 'ixemplee'})['value']
-        return True
-
-    def punch(self, option, remark=''):
-        res = self.session.post(urljoin(SITE_URL, 'punch3.php'),
-                                data={
-                                    'comp': self.company,
-                                    'name': self.employee,
-                                    'remark': remark,
-                                    'B1': option,
-                                    'ix': self.ix_employee,
-                                    'ts': '',
-                                    'allowremarks': 1,
-                                    'msgfound': 0,
-                                    'thetask': 0,
-                                    'teamleader': 0,
-                                    'tflag': ''
-                                }, headers=HEADERS)
-
-        if res.status_code != 200:
-            raise TimeWatchError(res.text)
-        return True
-
-    def punch_in(self, remark=''):
-        return self.punch(PUNCH_IN, remark)
-
-    def punch_out(self, remark=''):
-        return self.punch(PUNCH_OUT, remark)
